backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
waitress==2.1.2

